test: reduce unmanaged parallelism in domain test
The original test lauches 10 child processes at once
and bypass `test.py`'s process regulation.
This PR reduces the unmanaged parallelism and is a
temporary workaround for #9979 (not a real fix).

PR-URL: https://github.com/nodejs/node/pull/10329
