+++ test/parallel/test-domain-no-error-handler-abort-on-uncaught-4.js
@@ -1,0 +1,20 @@
+'use strict';
+
+const common = require('../common');
+const domain = require('domain');
+
+function test() {
+  const d = domain.create();
+
+  d.run(function() {
+    process.nextTick(function() {
+      throw new Error('boom!');
+    });
+  });
+}
+
+if (process.argv[2] === 'child') {
+  test();
+} else {
+  common.childShouldThrowAndAbort();
+}
