+++ test/parallel/test-domain-no-error-handler-abort-on-uncaught-9.js
@@ -1,0 +1,27 @@
+'use strict';
+
+const common = require('../common');
+const domain = require('domain');
+
+function test() {
+  const d = domain.create();
+  const d2 = domain.create();
+
+  d.on('error', function errorHandler() {
+  });
+
+  d.run(function() {
+    d2.run(function() {
+      var fs = require('fs');
+      fs.exists('/non/existing/file', function onExists() {
+        throw new Error('boom!');
+      });
+    });
+  });
+}
+
+if (process.argv[2] === 'child') {
+  test();
+} else {
+  common.childShouldThrowAndAbort();
+}
